+++ apps/account-server/src/context/state-machine-context.tsx
@@ -73,6 +73,33 @@ export const MainStateMachineContextProvider = ({
               },
             };
           }),
+          logout: assign(({ context }) => {
+            if (context.requests.incoming) {
+              const successResponse = {
+                id: crypto.randomUUID(),
+                requestId: context.requests.incoming.id,
+                content: {
+                  result: null,
+                },
+              };
+              windowService.sendMessage(successResponse);
+            }
+
+            return {
+              ...context,
+              isAuthenticated: false,
+              isLoadingResources: false,
+              requests: {
+                incoming: null,
+                typedDataSigning: null,
+                messageSigning: null,
+                transaction: null,
+                session: null,
+                authentication: null,
+                logout: null,
+              },
+            };
+          }),
           clearProfileRequests: assign(({ context }) => {
             if (context.requests.incoming) {
               const successResponse = {