+++ packages/account-react-native/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sophon-labs/account-react-native
 
+## 2.1.2
+
+### Patch Changes
+
+- Bugfixes and disconnect/logout support
+- Updated dependencies
+  - @sophon-labs/account-message-bridge@2.1.2
+  - @sophon-labs/account-communicator@2.1.2
+  - @sophon-labs/account-provider@2.1.2
+  - @sophon-labs/account-core@2.1.2
+
 ## 2.1.1
 
 ### Patch Changes